requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==1.11.8
django-cors-headers==2.1.0
django-redis==4.8.0
djangorestframework==3.7.7
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pycryptodome==3.4.11
Pygments==2.2.0
PyMySQL==0.8.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
